+++ src/main/kotlin/com/workoutperf/model/ContestPosition.kt
@@ -3,7 +3,7 @@ package com.workoutperf.model
 data class ContestPosition(
         override val id: String?,
         override val rank: Int,
-        override val points: Int,
+        override val points: Double,
         override var leaderboard: ContestLeaderboard?,
         val workoutPositions: MutableMap<Workout, WorkoutPosition>,
         var contest: Contest?,
