+++ src/main/kotlin/com/workoutperf/model/Position.kt
@@ -3,6 +3,6 @@ package com.workoutperf.model
 interface Position<L> {
     val id: String?
     val rank: Int
-    val points: Int
+    val points: Double
     var leaderboard: L?
 }
