+++ src/main/kotlin/com/workoutperf/model/WorkoutPosition.kt
@@ -3,7 +3,7 @@ package com.workoutperf.model
 data class WorkoutPosition(
         override val id: String?,
         override val rank: Int,
-        override val points: Int,
+        override val points: Double,
         override var leaderboard: WorkoutLeaderboard?,
         val performance: Performance?,
         val athlete: Person?,
